improve formatting and type hinting

+++ RegularConsumer.py
@@ -9,11 +9,13 @@ class RegularConsumer(Consumer):
 
     def GetMonthlyConsumption(self, month: int) -> float:
         month = int(np.ceil(month % 12)) if (month % 12) > 0 else 12
-        result = np.sum(hlp.SliceMonth(array=self.demandProfile, month=month))
+        result = np.sum(
+            hlp.SliceMonth(array=self.demandProfile, month=month), dtype=float
+        )
         return result
 
     def GetYearlyConsumption(self) -> float:
-        return np.sum(self.demandProfile)
+        return np.sum(self.demandProfile, dtype=float)
 
     def GetYearlyExpenditure(self, consumptionTariff: ElectricityTariff) -> float:
         demand = self.GetYearlyConsumption()
